refactor: split GoogleLoginButton component #20
